pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "confluence-wiki"
version = "0.1.0"
description = "Confluence Data Center mirror based markdown wiki service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "beautifulsoup4>=4.13.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.0",
  "langchain-openai>=0.3.0",
  "markdown-it-py>=3.0.0",
  "openai>=1.99.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "python-slugify>=8.0.0",
  "pyyaml>=6.0.0",
  "sqlalchemy>=2.0.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=1.0.0",
]

[project.scripts]
confluence-wiki = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
